+++ app/src/main/ui/src/pages/Account.tsx
@@ -1,0 +1,5 @@
+export default function Account() {
+  return <div>
+
+  </div>
+}
