+++ app/src/main/ui/src/pages/Groups.tsx
@@ -1,0 +1,3 @@
+export default function Groups() {
+  return <>Groups</>
+}
