+++ app/src/main/ui/src/pages/Users.tsx
@@ -1,0 +1,3 @@
+export default function Users() {
+  return <>Users</>
+}
